Newton-Raphson method for computing MLE under logistic regression model

For a simple logistic regression model with an intercept and one slope parameter, this function returns their respective MLE's. Iterations stop once the improvement in successive steps falls under a certain tolerance level.

NR_logit(x, y, n, tol = 1e-06, verbose = FALSE)

Arguments

x

vector of covariates (predictor)

y

vector of responses, each one distributed as a Binomial random variable

n

vector of number of trials for each observation.

tol

tolerance level governing when to stop the iterations. Here we use the larger absolute error for either parameter and compare it to tol. Iterations stop when the error is less than tol. Defaults to 1e-6.

verbose

Logical; if TRUE then function prints iterative feedback to the console, and if FALSE there is no printing. Defaults to FALSE.

Value

A data.frame with two entries in one row, the first being the MLE of the intercept and the second being the MLE of the slope.

Details

We can view y as the number of successes out of n trials, with probability of success p unknown. The linear predictor is related to p by the log odds ratio: \(log(p/(1-p)) = \alpha + \beta x\). It is known that in logsitic regression, we cannot obtain closed form solutions for \(\alpha\) and \(\beta\). This function uses the Newton-Raphson method to estimate these parameters. Our initial value is (\(\alpha_0\), \(\beta_0)\) = (0, 0), an intuitive choice that corresponds to the case of equally likely outcomes (p = 0.5).

In the function definition, the matrix L is the system of nonlinear score functions we need to solve, and L_prime is the derivative of L. Hence, each step of the iteration computes updated estimates via (\(\alpha_{i+1}\), \(\beta_{i+1}\)) = (\(\alpha_{i}\), \(\beta_{i}\)) + L_prime^-1 * L.

Examples

# Generate random covariates, responses, and trial totals
set.seed(547)
x <- runif(15, 6, 10)
n <- sample(1:15, replace = TRUE)
y <- rbinom(15, n, 0.6)
NR_logit(x, y, n)

#>   intercept      slope
#> 1  3.268903 -0.3770642

NR_logit(x, y, n, verbose = TRUE)

#> [1] "Initial Value: intercept = 0 slope = 0"
#> [1] "Iteration 1 : intercept = 3.100302 , slope = -0.357474"
#> [1] "Iteration 2 : intercept = 3.267689 , slope = -0.376923"
#> [1] "Iteration 3 : intercept = 3.268903 , slope = -0.377064"
#> [1] "Iteration 4 : intercept = 3.268903 , slope = -0.377064"
#> [1] "MLE are intercept = 3.268903 and slope = -0.377064"

#>   intercept      slope
#> 1  3.268903 -0.3770642

NR_logit(x, y, n, tol = 1e-10, verbose = TRUE)

#> [1] "Initial Value: intercept = 0 slope = 0"
#> [1] "Iteration 1 : intercept = 3.100302 , slope = -0.357474"
#> [1] "Iteration 2 : intercept = 3.267689 , slope = -0.376923"
#> [1] "Iteration 3 : intercept = 3.268903 , slope = -0.377064"
#> [1] "Iteration 4 : intercept = 3.268903 , slope = -0.377064"
#> [1] "Iteration 5 : intercept = 3.268903 , slope = -0.377064"
#> [1] "MLE are intercept = 3.268903 and slope = -0.377064"

#>   intercept      slope
#> 1  3.268903 -0.3770642